+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/sync/SyncConfiguration.kt
@@ -39,8 +39,8 @@ import io.realm.kotlin.mongodb.exceptions.ClientResetRequiredException
 import io.realm.kotlin.mongodb.exceptions.SyncException
 import io.realm.kotlin.mongodb.internal.SyncConfigurationImpl
 import io.realm.kotlin.mongodb.internal.UserImpl
+import io.realm.kotlin.types.BaseRealmObject
 import io.realm.kotlin.types.ObjectId
-import io.realm.kotlin.types.RealmObject
 import io.realm.kotlin.types.RealmUUID
 import kotlin.reflect.KClass
 import kotlin.time.Duration
@@ -210,7 +210,7 @@ public interface SyncConfiguration : Configuration {
     public class Builder private constructor(
         private var user: User,
         private var partitionValue: PartitionValue?,
-        schema: Set<KClass<out RealmObject>>,
+        schema: Set<KClass<out BaseRealmObject>>,
     ) : Configuration.SharedBuilder<SyncConfiguration, Builder>(schema) {
 
         // Shouldn't default to 'default.realm' - Object Store will generate it according to which
@@ -235,7 +235,7 @@ public interface SyncConfiguration : Configuration {
          */
         public constructor(
             user: User,
-            schema: Set<KClass<out RealmObject>>
+            schema: Set<KClass<out BaseRealmObject>>
         ) : this(user, null as PartitionValue?, schema)
 
         /**
@@ -255,7 +255,7 @@ public interface SyncConfiguration : Configuration {
         public constructor(
             user: User,
             partitionValue: ObjectId?,
-            schema: Set<KClass<out RealmObject>>
+            schema: Set<KClass<out BaseRealmObject>>
         ) : this(user, PartitionValue(partitionValue), schema)
 
         /**
@@ -292,7 +292,7 @@ public interface SyncConfiguration : Configuration {
         public constructor(
             user: User,
             partitionValue: Int?,
-            schema: Set<KClass<out RealmObject>>
+            schema: Set<KClass<out BaseRealmObject>>
         ) : this(user, PartitionValue(partitionValue), schema)
 
         /**
@@ -311,7 +311,7 @@ public interface SyncConfiguration : Configuration {
         public constructor(
             user: User,
             partitionValue: Long?,
-            schema: Set<KClass<out RealmObject>>
+            schema: Set<KClass<out BaseRealmObject>>
         ) : this(user, PartitionValue(partitionValue), schema)
 
         /**
@@ -330,7 +330,7 @@ public interface SyncConfiguration : Configuration {
         public constructor(
             user: User,
             partitionValue: String?,
-            schema: Set<KClass<out RealmObject>>
+            schema: Set<KClass<out BaseRealmObject>>
         ) : this(user, PartitionValue(partitionValue), schema)
 
         init {
@@ -608,7 +608,7 @@ public interface SyncConfiguration : Configuration {
          * literals.
          * @throws IllegalArgumentException if the user is not valid and logged in.
          */
-        public fun create(user: User, schema: Set<KClass<out RealmObject>>): SyncConfiguration =
+        public fun create(user: User, schema: Set<KClass<out BaseRealmObject>>): SyncConfiguration =
             Builder(user, schema).build()
 
         /**
@@ -626,7 +626,7 @@ public interface SyncConfiguration : Configuration {
         public fun create(
             user: User,
             partitionValue: String?,
-            schema: Set<KClass<out RealmObject>>
+            schema: Set<KClass<out BaseRealmObject>>
         ): SyncConfiguration =
             Builder(user, partitionValue, schema).build()
 
@@ -645,7 +645,7 @@ public interface SyncConfiguration : Configuration {
         public fun create(
             user: User,
             partitionValue: Int?,
-            schema: Set<KClass<out RealmObject>>
+            schema: Set<KClass<out BaseRealmObject>>
         ): SyncConfiguration =
             Builder(user, partitionValue, schema).build()
 
@@ -664,7 +664,7 @@ public interface SyncConfiguration : Configuration {
         public fun create(
             user: User,
             partitionValue: Long?,
-            schema: Set<KClass<out RealmObject>>
+            schema: Set<KClass<out BaseRealmObject>>
         ): SyncConfiguration =
             Builder(user, partitionValue, schema).build()
 
@@ -678,7 +678,7 @@ public interface SyncConfiguration : Configuration {
          * @throws IllegalArgumentException if the user is not valid and logged in.
          * * **See:** [partition key](https://www.mongodb.com/docs/realm/sync/data-access-patterns/partitions/)
          */
-        public fun create(user: User, partitionValue: ObjectId?, schema: Set<KClass<out RealmObject>>): SyncConfiguration =
+        public fun create(user: User, partitionValue: ObjectId?, schema: Set<KClass<out BaseRealmObject>>): SyncConfiguration =
             Builder(user, partitionValue, schema).build()
 
         /**